+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Volumes settings based on hardware volumes. Remove software gains.
 - Display call duration instead of conference duration.
 - Publish presence indicator synchronized on default account.
+- Remove from GUI a bad account created from Assistant.
 
 ### Added
 - Screen Sharing in conference calls